Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
pydantic>=2.11,<3
pydantic-settings==2.4.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
psycopg[binary]>=3.1,<4
httpx==0.27.2
beautifulsoup4==4.12.3

